pyproject.toml
[project]
name = "mnemodyn-draft"
version = "0.1.0"
description = "MnemoDyn draft environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "absl-py==2.1.0",
  "aiohappyeyeballs==2.5.0",
  "aiohttp==3.11.13",
  "aiosignal==1.3.2",
  "asttokens==3.0.0",
  "attrs==25.1.0",
  "certifi==2025.1.31",
  "charset-normalizer==3.4.1",
  "colorama==0.4.6",
  "comm==0.2.2",
  "contourpy==1.3.1",
  "cycler==0.12.1",
  "debugpy==1.8.13",
  "decorator==5.2.1",
  "einops==0.8.1",
  "exceptiongroup==1.2.2",
  "executing==2.1.0",
  "filelock==3.17.0",
  "fonttools==4.56.0",
  "frozenlist==1.5.0",
  "fsspec==2025.3.0",
  "grpcio==1.71.0",
  "idna==3.10",
  "importlib_metadata==8.6.1",
  "importlib_resources==6.5.2",
  "ipykernel==6.29.5",
  "ipython==9.0.2",
  "ipython_pygments_lexers==1.1.1",
  "jedi==0.19.2",
  "Jinja2==3.1.6",
  "joblib==1.4.2",
  "jupyter_client==8.6.3",
  "jupyter_core==5.7.2",
  "kiwisolver==1.4.8",
  "lightning==2.5.0.post0",
  "lightning-utilities==0.14.0",
  "Markdown==3.7",
  "MarkupSafe==3.0.2",
  "matplotlib==3.10.1",
  "matplotlib-inline==0.1.7",
  "mpmath==1.3.0",
  "multidict==6.1.0",
  "nest_asyncio==1.6.0",
  "networkx==3.4.2",
  "nibabel==5.3.2",
  "numpy==2.2.3",
  "nvidia-cublas-cu12==12.4.5.8",
  "nvidia-cuda-cupti-cu12==12.4.127",
  "nvidia-cuda-nvrtc-cu12==12.4.127",
  "nvidia-cuda-runtime-cu12==12.4.127",
  "nvidia-cudnn-cu12==9.1.0.70",
  "nvidia-cufft-cu12==11.2.1.3",
  "nvidia-curand-cu12==10.3.5.147",
  "nvidia-cusolver-cu12==11.6.1.9",
  "nvidia-cusparse-cu12==12.3.1.170",
  "nvidia-cusparselt-cu12==0.6.2",
  "nvidia-nccl-cu12==2.21.5",
  "nvidia-nvjitlink-cu12==12.4.127",
  "nvidia-nvtx-cu12==12.4.127",
  "packaging==24.2",
  "pandas==2.2.3",
  "parso==0.8.4",
  "pexpect==4.9.0",
  "pickleshare==0.7.5",
  "pillow==11.1.0",
  "pip==25.0",
  "platformdirs==4.3.6",
  "prompt_toolkit==3.0.50",
  "propcache==0.3.0",
  "protobuf==6.30.0",
  "psutil==7.0.0",
  "ptyprocess==0.7.0",
  "pure_eval==0.2.3",
  "Pygments==2.19.1",
  "pyparsing==3.2.1",
  "python-dateutil==2.9.0.post0",
  "pytorch-lightning==2.5.0.post0",
  "pytorch_wavelets==1.3.0",
  "pytz==2025.1",
  "PyWavelets==1.8.0",
  "PyYAML==6.0.2",
  "pyzmq==26.3.0",
  "requests==2.32.3",
  "scikit-learn==1.6.1",
  "scipy==1.15.2",
  "setuptools==75.8.0",
  "six==1.17.0",
  "stack_data==0.6.3",
  "sympy==1.13.1",
  "tabulate==0.9.0",
  "tensorboard==2.19.0",
  "tensorboard-data-server==0.7.2",
  "tensorboardX==2.6.2.2",
  "threadpoolctl==3.5.0",
  "torch==2.6.0",
  "torch-tb-profiler==0.4.3",
  "torchcde==0.2.5",
  "torchdiffeq==0.2.5",
  "torchmetrics==1.6.2",
  "torchsde==0.2.6",
  "tornado==6.4.2",
  "tqdm==4.67.1",
  "traitlets==5.14.3",
  "trampoline==0.1.2",
  "triton==3.2.0",
  "typing_extensions==4.12.2",
  "tzdata==2025.1",
  "urllib3==2.3.0",
  "wcwidth==0.2.13",
  "Werkzeug==3.1.3",
  "wheel==0.45.1",
  "yarl==1.18.3",
  "zipp==3.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
